containers/ai-services/vision/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.4
requests==2.31.0
aiofiles==23.2.1
pydantic==2.5.0
structlog==23.2.0
prometheus-client==0.19.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1